Fix adding directories logic and add check function
Adding an entire directory in the database and keeping track only of one SHA256 turned out to be a bad idea, as changing a directory's contents and changing them back again wouldn't result in the same hash as before. This may be due to the order the files are recursively parsed.

Now, when adding a directory, its files are added instead, with a SHA256 for each file. This won't detect new foreign files, but it doesn't seem that useful do detect this either.

Also, now there's a function to check for the integrity of all the files in the database.